TG-76: admins and employees controllers firebase implement

+++ src/controllers/admins.js
@@ -1,6 +1,5 @@
 import mongoose from 'mongoose';
 import firebase from '../helpers/firebase';
-// import Users from '../models/Users';
 import Admins from '../models/Admins';
 
 const getAllAdmins = async (req, res) => {
@@ -128,18 +127,19 @@ const updateAdmins = async (req, res) => {
       { new: true },
     );
 
-    await firebase.auth().updateUser(req.body.firebaseUid, {
-      email: req.body.email,
-      password: req.body.password,
-    });
-
     if (updatedAdmin) {
+      await firebase.auth().updateUser(req.body.firebaseUid, {
+        email: req.body.email,
+        password: req.body.password,
+      });
+
       return res.status(200).json({
         message: `Admin with id ${req.params.id} updated successfully`,
         data: updatedAdmin,
         error: false,
       });
     }
+
     return res.status(404).json({
       message: `Admin with id ${req.params.id} not found`,
       data: undefined,
@@ -164,13 +164,15 @@ const deleteAdmins = async (req, res) => {
   }
   try {
     const findAdminById = await Admins.findById(req.params.id);
-    firebase.auth().deleteUser(findAdminById.firebaseUid);
+    if (findAdminById) {
+      await firebase.auth().deleteUser(findAdminById.firebaseUid);
+      const deletedAdmin = await Admins.findByIdAndDelete(req.params.id);
 
-    const deletedAdmin = await Admins.findByIdAndDelete(req.params.id);
-
-    if (deletedAdmin) {
-      return res.status(204).json();
+      if (deletedAdmin) {
+        return res.status(204).json();
+      }
     }
+
     return res.status(404).json({
       message: `Admin with id ${req.params.id} not found`,
       data: undefined,
