+++ src/controllers/employees.js
@@ -1,5 +1,6 @@
 import mongoose from 'mongoose';
 import Employees from '../models/Employees';
+import firebase from '../helpers/firebase';
 
 const getAllEmployees = async (req, res) => {
   try {
@@ -82,9 +83,16 @@ const createEmployee = async (req, res) => {
       lastName: req.body.lastName,
       phone: req.body.phone,
       email: req.body.email,
-      password: req.body.password,
       dni: req.body.dni,
     });
+    const newFirebaseUser = await firebase.auth().createUser({
+      email: req.body.email,
+      password: req.body.password,
+    });
+
+    await firebase
+      .auth()
+      .setCustomUserClaims(newFirebaseUser.uid, { role: 'EMPLOYEE' });
 
     const result = await newEmployee.save();
     return res.status(201).json({
@@ -119,6 +127,11 @@ const updateEmployee = async (req, res) => {
       { new: true },
     );
     if (updatedEmployee) {
+      await firebase.auth().updateUser(req.body.firebaseUid, {
+        email: req.body.email,
+        password: req.body.password,
+      });
+
       return res.status(200).json({
         message: `Employee with id ${id} updated successfully`,
         data: updatedEmployee,
@@ -151,15 +164,21 @@ const deleteEmployee = async (req, res) => {
   }
   try {
     const { id } = req.params;
-    const deletedEmployee = await Employees.findByIdAndDelete(id);
+    const findEmployeesById = await Employees.findById(req.params.id);
 
-    if (deletedEmployee) {
-      return res.status(200).json({
-        message: 'Employee deleted succesfully',
-        data: deletedEmployee,
-        error: false,
-      });
+    if (findEmployeesById) {
+      firebase.auth().deleteUser(findEmployeesById.firebaseUid);
+      const deletedEmployee = await Employees.findByIdAndDelete(id);
+
+      if (deletedEmployee) {
+        return res.status(200).json({
+          message: 'Employee deleted succesfully',
+          data: deletedEmployee,
+          error: false,
+        });
+      }
     }
+
     return res.status(404).json({
       message: `Employee with id ${id} not found`,
       data: undefined,
